scripts/python/requirements.txt
# SITES Spectral Python Scripts Requirements
# Install with: pip install -r requirements.txt
# Or with uv: uv pip install -r requirements.txt

requests>=2.31.0
python-dotenv>=1.0.0
